+++ tests/tmp_store.c
@@ -7,7 +7,7 @@
 #define STORE_PREFIX "/tmp/oqs-temp-file-"
 #define MAXPATHLEN 64
 
-static OQS_STATUS oqs_fstore(const char *fname, uint8_t *data, int len) {
+static OQS_STATUS oqs_fstore(const char *fname, uint8_t *data, size_t len) {
 	char fpath[MAXPATHLEN];
 	strcpy(fpath, STORE_PREFIX);
 	FILE *fp = fopen(strcat(fpath, fname), "wb");
@@ -20,7 +20,7 @@ static OQS_STATUS oqs_fstore(const char *fname, uint8_t *data, int len) {
 	return OQS_SUCCESS;
 }
 
-static OQS_STATUS oqs_fload(const char *fname, uint8_t *data, int len) {
+static OQS_STATUS oqs_fload(const char *fname, uint8_t *data, size_t len) {
 	char fpath[MAXPATHLEN];
 	strcpy(fpath, STORE_PREFIX);
 	FILE *fp = fopen(strcat(fpath, fname), "r");
@@ -29,7 +29,7 @@ static OQS_STATUS oqs_fload(const char *fname, uint8_t *data, int len) {
 		return OQS_ERROR;
 	}
 	if (fread(data, len, 1, fp) != 1) {
-		fprintf(stderr, "Couldn't read all %d bytes correctly (operations called in proper sequence?). Exiting.\n", len);
+		fprintf(stderr, "Couldn't read all %ld bytes correctly (operations called in proper sequence?). Exiting.\n", len);
 		return OQS_ERROR;
 	}
 	fclose(fp);
